requirements.txt
tensorflow==2.15.0
librosa==0.10.1
numpy==1.26.4
scipy==1.12.0
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
pandas==2.2.1
pyyaml==6.0.1
h5py==3.10.0  # For saving/loading .h5 models if needed
hydra-core==1.3.2  # If using Hydra for configs (optional, based on earlier suggestions)
